+++ src/components/index.tsx
@@ -16,6 +16,9 @@ export { default as DeleteConfirmAction } from './DeleteConfirmAction';
 export type { FloatDrawerProps } from './FloatDrawer';
 export { default as FloatDrawer } from './FloatDrawer';
 
+export type { LoadingProps } from './Loading';
+export { default as Loading } from './Loading';
+
 export type { ModalActionProps, FormCompPropsConstraint, ModalActionTrigger, ModalActionRef } from './ModalAction';
 export { withDefaultModalActionProps, withModalAction } from './ModalAction';
 export { default as ModalAction } from './ModalAction';